requirements.txt
absl-py==1.4.0
adabelief-pytorch==0.2.1
aiosignal==1.3.1
alabaster==0.7.12
altair==4.2.2
ansi2html==1.8.0
anyio==3.5.0
appdirs==1.4.4
argh==0.28.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
astroid==2.11.7
astropy==5.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.5.1
bkcharts==0.2
black==22.6.0
bleach==4.1.0
blinker==1.6.2
bokeh==2.4.3
boto3==1.24.28
botocore==1.27.28
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
category-encoders==2.6.1
certifi==2023.11.17
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.4
cloudpickle==2.2.1
clyent==1.2.2
cmdstanpy==1.2.4
colorama==0.4.5
colorcet==3.0.0
comtypes==1.1.10
constantly==15.1.0
contourpy==1.1.0
cookiecutter==1.7.3
cryptography==37.0.1
cssselect==1.1.0
cvxopt==1.3.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.0
daal4py==2021.6.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.7.0
datashader==0.14.1
datashape==0.5.4
deap==1.3.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diff-match-patch==20200713
dill==0.3.4
distlib==0.3.6
distributed==2022.7.0
dm-tree==0.1.8
dnspython==2.3.0
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
faiss-cpu==1.7.4
Farama-Notifications==0.0.4
fastjsonschema==2.16.2
filelock==3.12.3
flake8==4.0.1
Flask==1.1.2
flatbuffers==23.3.3
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.12.2
funcy==1.17
future==0.18.2
gast==0.4.0
gensim==4.1.2
gitdb==4.0.10
GitPython==3.1.31
glob2==0.7
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==1.1.1
grpcio==1.53.0
gurobipy==11.0.1
gym==0.26.2
gym-notices==0.0.8
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h11==0.14.0
h5py==3.7.0
HeapDict==1.0.1
holidays==0.52
holoviews==1.15.0
huggingface-hub==0.16.4
hvplot==0.8.0
hyperlink==21.0.0
idna==3.6
imagecodecs==2021.8.26
imageio==2.19.3
imagesize==1.4.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.6.0
importlib_resources==6.1.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
intake==0.6.5
intervaltree==3.1.0
ipykernel==6.15.2
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.0.1
jax==0.4.8
jax-jumpy==1.0.0
jdcal==1.4.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.2.0
JPype1==1.4.1
json5==0.9.6
jsonschema==4.16.0
kaleido==0.2.1
keras==2.11.0
keyring==23.4.0
kiwisolver==1.4.5
konlpy==0.6.0
lazy-object-proxy==1.6.0
libarchive-c==2.9
libclang==16.0.0
lightgbm==3.3.5
llvmlite==0.38.0
locket==1.0.0
lxml==4.9.1
lz4==3.1.3
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
menuinst==1.4.19
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.0.4
mock==4.0.3
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.3
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.7
nose==1.3.7
notebook==6.4.12
numba==0.55.1
numexpr==2.8.3
numpy==1.21.2
numpydoc==1.4.0
oauthlib==3.2.2
olefile==0.46
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.1
outcome==1.2.0
packaging==23.1
pandas==1.3.3
pandocfilters==1.5.0
panel==0.13.1
param==1.12.0
paramiko==2.8.1
parsel==1.6.0
parso==0.8.3
partd==1.2.0
pathlib==1.0.1
pathspec==0.9.0
patsy==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==24.1.2
pkginfo==1.8.2
platformdirs==2.5.2
plotly==5.16.1
plotly-resampler==0.8.3.2
pluggy==1.0.0
pmdarima==2.0.3
poyo==0.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
prophet==1.1.5
Protego==0.1.16
protobuf==3.19.6
psutil==5.9.0
psycopg2==2.9.9
ptyprocess==0.7.0
pvlib==0.10.2
py==1.11.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycaret==3.0.2
pyclustering==0.10.1.2
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.21
pyct==0.4.8
pycurl==7.45.1
pydeck==0.8.1b0
PyDispatcher==2.0.5
pydocstyle==6.1.1
pyerfa==2.0.0
pyflakes==2.4.0
pygame==2.1.0
PyGetWindow==0.0.9
pyglet==1.2.4
Pygments==2.15.0
PyHamcrest==2.0.2
PyJWT==2.4.0
pyLDAvis==3.3.1
pylint==2.14.5
pyls-spyder==0.4.0
pymongo==4.3.3
Pympler==1.0.1
PyMsgBox==1.0.9
PyMySQL==1.1.0
PyNaCl==1.5.0
pynndescent==0.5.10
pyod==1.0.9
pyodbc==4.0.34
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.0
PyScreeze==0.1.28
pyserial==3.5
PySocks==1.7.1
pyswarm==0.6
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-pptx==1.0.2
python-slugify==5.0.2
python-snappy==0.6.0
pytils==0.4.1
pytweening==1.0.4
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.0.2
PyWavelets==1.3.0
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==6.0.1
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.2.2
QtPy==2.2.0
queuelib==1.5.0
ray==2.3.1
regex==2022.7.9
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.3.3
rope==0.22.0
rsa==4.9
Rtree==0.9.7
ruamel-yaml-conda==0.15.100
s3transfer==0.6.0
safetensors==0.3.1
schemdraw==0.15
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-learn-intelex==2021.20221004.171935
scikit-plot==0.3.7
scipy==1.10.1
Scrapy==2.6.2
seaborn==0.12.2
selenium==4.7.2
semver==3.0.1
Send2Trash==1.8.0
service-identity==18.1.0
setuptools==63.4.1
sip==4.19.13
six==1.16.0
sklearn==0.0.post1
sktime==0.17.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.2.2
spyder-kernels==2.2.1
SQLAlchemy==2.0.21
stanio==0.5.1
statsmodels==0.13.2
streamlit==1.20.0
streamlit-option-menu==0.3.2
sumolib==1.20.0
svgwrite==1.4.3
swig==4.1.1
sympy==1.12
tables==3.6.1
tabulate==0.8.10
tai==0.0.45
tbats==1.1.3
TBB==0.2
tblib==1.7.0
tdqm==0.0.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.13.1
testpath==0.6.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
tldextract==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
too==0.0.1
tool==0.8.0
tools==0.1.9
toolz==0.11.2
torch==2.1.2
torchaudio==2.0.2
torchvision==0.16.2
tornado==6.1
tqdm==4.64.1
trace-updater==0.0.9.1
traci==1.20.0
traitlets==5.1.1
transformers==4.31.0
Tree==0.2.4
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txtai==5.5.1
typer==0.7.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3
ujson==5.4.0
umap-learn==0.5.3
Unidecode==1.2.0
urllib3==2.1.0
validators==0.20.0
virtualenv==20.21.0
w3lib==1.21.0
watchdog==2.1.6
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
windows-curses==2.3.1
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.2.0
xarray==0.20.1
xgboost==1.7.5
xlrd==2.0.1
XlsxWriter==3.0.3
xlwings==0.27.15
xxhash==3.2.0
yapf==0.31.0
yellowbrick==1.5
zict==2.1.0
zipp==3.8.0
zope.interface==5.4.0
